project_ban/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-ban"
version = "1.0.0"
description = "Bare-metal Autonomous Node — strips macOS to bare-metal over SSH, launches priority workload"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "BAN Team"}]
dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.0.0",
    "paramiko>=3.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.scripts]
ban = "src.cli:app"

[tool.setuptools.packages.find]
include = ["src*"]
